+++ Navigation/Profile/ProfileViewController.swift
@@ -8,45 +8,112 @@ import UIKit
 
 class ProfileViewController: UIViewController {
   
-    let profileHeaderView = ProfileHeaderView()
+    //let profileHeaderView = ProfileHeaderView()
     let profileTableHeaderView = ProfileTableHeaderView()
+    
+    // MARK: - Data
+    
+    let data = PostModel.make()
+    
+    private enum CellReuseID: String {
+        case base = "BaseTableViewCell_ReuseID"
+        case custom = "CustomTableViewCell_ReuseID"
+    }
+    
+    private enum HeaderFooterReuseID: String {
+        case base = "TableSelectionFooterHeaderView_ReuseID"
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
         view.backgroundColor = .lightGray
         title = "Profile" 
-        view.addSubview(profileHeaderView)
+        //view.addSubview(profileHeaderView)
+        view.addSubview(profileTableHeaderView)
         setupConstraints()
     }
  
     func setupConstraints() {
         let safeAreaGuide = view.safeAreaLayoutGuide
-        profileHeaderView.translatesAutoresizingMaskIntoConstraints = false
+        profileTableHeaderView.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
             
-            profileHeaderView.topAnchor.constraint(
+            profileTableHeaderView.topAnchor.constraint(
                 equalTo: safeAreaGuide.topAnchor,
                 constant: 10
             ),
             
-            profileHeaderView.bottomAnchor.constraint(
+            profileTableHeaderView.bottomAnchor.constraint(
                 equalTo: view.bottomAnchor,
                 constant: 10
             ),
             
-            profileHeaderView.leadingAnchor.constraint(
+            profileTableHeaderView.leadingAnchor.constraint(
                 equalTo: view.leadingAnchor,
                 constant: 0
             ),
-            profileHeaderView.trailingAnchor.constraint(
+            profileTableHeaderView.trailingAnchor.constraint(
                 equalTo: view.trailingAnchor,
                 constant: 0
             ),
             
-            profileHeaderView.heightAnchor.constraint(equalTo: view.heightAnchor, constant: 220)
+            profileTableHeaderView.heightAnchor.constraint(equalTo: view.heightAnchor, constant: 220)
         ])
     }
      
 }
 
+extension ProfileViewController: UITableViewDataSource, UITableViewDelegate {
+    
+    func tableView(
+        _ tableView: UITableView,
+        numberOfRowsInSection section: Int
+    ) -> Int {
+        data.count
+    }
+    
+    func tableView(
+        _ tableView: UITableView,
+        cellForRowAt indexPath: IndexPath
+    ) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(
+            withIdentifier: CellReuseID.base.rawValue,
+            for: indexPath
+        ) as? PostTableViewCell else {
+            fatalError("could not dequeueReusableCell")
+        }
+        cell.update(data[indexPath.row])
+        return cell
+    }
+    
+    func tableView(
+        _ tableView: UITableView,
+        heightForHeaderInSection section: Int
+    ) -> CGFloat {
+        UITableView.automaticDimension
+    }
+    
+    func tableView(
+        _ tableView: UITableView,
+        heightForFooterInSection section: Int
+    ) -> CGFloat {
+        UITableView.automaticDimension
+    }
+/*
+    func tableView(
+        _ tableView: UITableView,
+        viewForFooterInSection section: Int
+    ) -> UIView? {
+        
+        guard let footerView = tableView.dequeueReusableHeaderFooterView(
+            withIdentifier: HeaderFooterReuseID.base.rawValue
+        ) as? ProfileTableHeaderView else {
+            fatalError("could not dequeueReusableCell")
+        }
+        
+        footerView.update(title: "?")
+        return footerView
+    }
+*/
+}
